requirements.txt
Django>=3.2.4,<4.0.7
djangorestframework>=3.12.4,<3.13
psycopg2>=2.8
drf-spectacular>=0.15.1,<0.16
pytz==2021.3
